Moved old disc module and eaclog to disc package
Avoids name conflict and confusion of picard.discid with the discid pip package.
